+++ src/components/ids-area-chart/ids-area-chart.js
@@ -1,3 +1,4 @@
+import { attributes } from '../../core/ids-attributes';
 import { customElement, scss } from '../../core/ids-decorators';
 import Base from './ids-area-chart-base';
 import styles from './ids-area-chart.scss';
@@ -7,7 +8,10 @@ import styles from './ids-area-chart.scss';
  * @type {IdsAreaChart}
  * @inherits IdsElement
  * @mixes IdsEventsMixin
- * @part container - the outside container element
+ * @part svg - the outside svg element
+ * @part marker - the dots/markers in the chart
+ * @part line - the lines in the chart
+ * @part area - each area element in the chart
  */
 @customElement('ids-area-chart')
 @scss(styles)
@@ -39,12 +43,33 @@ export default class IdsAreaChart extends Base {
    * @private
    */
   #areas() {
-    let areas = '';
-    this.markerData.points.forEach((point, index) => {
-      if (this.markerData.points[index + 1]) {
-        areas += `M${point.left},${point.top}L${point.left},${this.markerData.gridBottom}L${this.markerData.points[index + 1]?.left},${this.markerData.gridBottom}L${this.markerData.points[index + 1]?.left},${this.markerData.points[index + 1]?.top}`;
-      }
+    let areaHTML = '';
+    this.markerData.points.forEach((pointGroup, groupIndex) => {
+      let areas = '';
+      pointGroup.forEach((point, index) => {
+        if (pointGroup[index + 1]) {
+          areas += `M${point.left},${point.top}L${point.left},${this.markerData.gridBottom}L${pointGroup[index + 1]?.left},${this.markerData.gridBottom}L${pointGroup[index + 1]?.left},${pointGroup[index + 1]?.top}`;
+        }
+      });
+      areaHTML += `<path part="area" d="${areas}Z" fill="var(${this.color(groupIndex)})"}></path>`;
     });
-    return `<path d="${areas}Z"></path>`;
+    return areaHTML;
+  }
+
+  /**
+   * Set the size of the markers (aka dots/ticks) in the chart
+   * @param {number} value The value to use (in pixels)
+   */
+  set markerSize(value) {
+    this.setAttribute(attributes.MARKER_SIZE, value);
+    this.rerender();
+  }
+
+  /**
+   * Adjust the size of the default marker
+   * @returns {number} value The value to use (in pixels)
+   */
+  get markerSize() {
+    return parseFloat(this.getAttribute(attributes.MARKER_SIZE)) || 1;
   }
 }