+++ src/components/ids-axis-chart/ids-chart-colors.js
@@ -1,0 +1,27 @@
+// Colors for showing different data sets
+const QUALITATIVE_COLORS = [
+  '--ids-color-palette-azure-80',
+  '--ids-color-palette-turquoise-40',
+  '--ids-color-palette-amethyst-60',
+  '--ids-color-palette-slate-40',
+  '--ids-color-palette-amber-30',
+  '--ids-color-palette-amber-70',
+  '--ids-color-palette-emerald-80',
+  '--ids-color-palette-azure-30',
+  '--ids-color-palette-ruby-80',
+  '--ids-color-palette-slate-20',
+  '--ids-color-palette-turquoise-60',
+  '--ids-color-palette-amber-90',
+  '--ids-color-palette-amethyst-30',
+  '--ids-color-palette-azure-60',
+  '--ids-color-palette-ruby-30',
+  '--ids-color-palette-amethyst-80',
+  '--ids-color-palette-emerald-30',
+  '--ids-color-palette-turquoise-80',
+  '--ids-color-palette-slate-60',
+  '--ids-color-palette-ruby-40'
+];
+
+export {
+  QUALITATIVE_COLORS
+};