+++ test/ids-area-chart/ids-area-chart-func-test.js
@@ -1,14 +1,14 @@
 /**
  * @jest-environment jsdom
  */
-import IdsLineChart from '../../src/components/ids-line-chart/ids-line-chart';
+import IdsAreaChart from '../../src/components/ids-area-chart/ids-area-chart';
 import dataset from '../../demos/data/components.json';
 
-describe('IdsLineChart Component', () => {
+describe('IdsAreaChart Component', () => {
   let lineChart;
 
   beforeEach(async () => {
-    lineChart = new IdsLineChart();
+    lineChart = new IdsAreaChart();
     document.body.appendChild(lineChart);
     lineChart.data = dataset;
   });
@@ -21,18 +21,11 @@ describe('IdsLineChart Component', () => {
     const errors = jest.spyOn(global.console, 'error');
 
     document.body.innerHTML = '';
-    lineChart = new IdsLineChart();
+    lineChart = new IdsAreaChart();
     document.body.appendChild(lineChart);
     lineChart.data = dataset;
 
     lineChart.remove();
     expect(errors).not.toHaveBeenCalled();
-  });
-
-  it('supports setting markerSize', () => {
-    expect(lineChart.markerSize).toEqual(5);
-    expect(lineChart.shadowRoot.querySelector('circle').getAttribute('r')).toEqual('5');
-    lineChart.markerSize = 8;
-    expect(lineChart.markerSize).toEqual(8);
   });
 });