+++ test/ids-axis-chart/ids-axis-chart-e2e-test.js
@@ -1,0 +1,17 @@
+describe('Ids Area Chart e2e Tests', () => {
+  const url = 'http://localhost:4444/ids-axis-chart';
+
+  beforeAll(async () => {
+    await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
+  });
+
+  it('should not have errors', async () => {
+    await expect(page.title()).resolves.toMatch('IDS Axis Chart Component');
+  });
+
+  it('should pass Axe accessibility tests', async () => {
+    await page.setBypassCSP(true);
+    await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
+    await expect(page).toPassAxeTests({ disabledRules: ['color-contrast'] });
+  });
+});